% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmpp_model_spec.R
\name{arg_funs}
\alias{arg_funs}
\alias{mmpp_model}
\alias{mmpp_control}
\title{Define specifics for model structure and fitting control}
\usage{
mmpp_model(form = ~1, link = "soft_plus", a = 1, L = 0, U = 0)

mmpp_control(
  lambda = mmpp_model(),
  q_r = mmpp_model(),
  q_m = mmpp_model(),
  struc = "mult",
  norm = TRUE
)
}
\arguments{
\item{form}{Model formula for MMPP model components, residency, movement, and detection}

\item{link}{A link function for the model component. One of \code{"soft_plus"}, \code{"log"}, or \code{"logit"}.}

\item{a}{Scale parameter for the \code{"soft_plus"} link function. Ignored for \code{link="log"} or \code{"logit"}. The \code{a} parameter determines the approximation to
a hard plus function, i.e., as \code{a} becomes large the soft plus function converges to \code{g^{-1}(x) = max(0,x)}. For this specification, \code{a} must be greater than or equal to 1.}

\item{L}{Lover bound for general logit link.}

\item{U}{Upper bound for general logit link.}

\item{lambda}{A named list for the detection portion of the model. See \verb{\link{mmpp_model}}}

\item{q_r}{A named list for the residency model. Must contain elements: \code{form}, \code{link}, and \code{a}.
The easiest way to construct this is the \verb{\link{mmpp_model}} function.}

\item{q_m}{A named list for the movement portion of the model. See \code{q_r} and \verb{\link{mmpp_model}}.}

\item{struc}{The form of the rate matrix entries. Can be one of: \code{"mult"} for residency times movement, i.e., q_{ij} = q_r(i) * pi_m(i,j) or \code{"add"} for a model of
the form q_{ij} = q_r(i) + pi_m(i,j).}

\item{norm}{Should the movement portion be adjusted to sum to 1. This is the parameterization suggested by Hewitt et al. (2023).}
}
\description{
Define specifics for model structure and fitting control
}
\references{
Hewitt, J., Gelfand, A. E., & Schick, R. S. (2023). Time-discretization approximation enriches continuous-time discrete-space models for animal movement. The Annals of Applied Statistics, 17:740-760.
}
