% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix.R
\name{dm_matrix}
\alias{dm_matrix}
\alias{dm_lambda}
\alias{dm_q}
\title{Create sparse representations of design matrices for
movement and resight models}
\usage{
dm_lambda(formula, ddl)

dm_q(formula, ddl)
}
\arguments{
\item{formula}{Model formula}

\item{ddl}{Design data list}
}
\description{
Create sparse representations of design matrices for
movement and resight models
}
\details{
This function is not designed for end-users but is exported for posterity.
}
