% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmpp.R, R/fit_mmpp_dir.R
\name{fit_mmpp_dir}
\alias{fit_mmpp_dir}
\title{Fit MMPP movement model to resight data}
\usage{
fit_mmpp_dir(
  data,
  ddl,
  model_parameters = list(lambda = list(form = ~1, offset = ~NULL), q = list(form = ~1,
    offset = ~log(1/num_neigh) - 1)),
  hessian = TRUE,
  start = NULL,
  method = "nlminb",
  fit = TRUE,
  debug = 0,
  ...
)

fit_mmpp_dir(
  data,
  ddl,
  model_parameters = list(lambda = list(form = ~1, offset = ~NULL), q = list(form = ~1,
    offset = ~log(1/num_neigh) - 1)),
  hessian = TRUE,
  start = NULL,
  method = "nlminb",
  fit = TRUE,
  debug = 0,
  ...
)
}
\arguments{
\item{data}{A processed data frame produced by the function \code{\link{process_data}}}

\item{ddl}{A design data list produced by the function \code{\link{make_design_data}}.}

\item{model_parameters}{Model formula for the detection and movement portions
of the MMPP model.}

\item{hessian}{Logical. Should the Hessian matrix be calculated to obtain the parameter
variance-covariance matrix.}

\item{start}{Optional starting values for the parameter must be a list of the
form \code{list(beta_l=c(), beta_q=c())}.}

\item{method}{Optimization method. See \code{\link[optimx]{optimr}}}

\item{fit}{Logical. Should the likelihood be optimized?}

\item{debug}{Integer from 1-4. Opens browser() at various points in the function call. Mostly for
package developers.}

\item{...}{Additional arguments passed to the optimization function
\code{\link[optimx]{optimr}} from the \code{\link[optimx]{optimx-package}}.}
}
\description{
Fit MMPP movement model to resight data

Fit MMPP movement model to resight data
}
\details{
Two model forms are available \code{list(lambda=list(form=~1, offset=NULL), q=list(form=~1, offset=~log(1.num_neigh)))}. For the
\code{q} model you must use \code{offset=0} to not have one. If it is left off, \code{offset=~log(1.num_neigh)))}
will be used. To use the movement rate form of Hewitt et al. (2023), one must use, e.g.,
\code{q = list(res_form = ~from_var, mov_form=~to_var, offset=...)}, where \code{from_var} is a variable
}
\references{
Hewitt, J., Gelfand, A. E., & Schick, R. S. (2023). Time-discretization approximation enriches continuous-time discrete-space models for animal movement. The Annals of Applied Statistics, 17:740-760.
}
\author{
Devin S. Johnson
}
