% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmpp_wmb.R
\name{fit_mmpp_wmb}
\alias{fit_mmpp_wmb}
\title{Fit MMPP movement model to resight data using a batch approximation to ML}
\usage{
fit_mmpp_wmb(
  data,
  ddl,
  model_parameters = list(lambda = list(form = ~1, offset = ~NULL), q = list(form = ~1,
    offset = ~log(1/num_neigh) - 1)),
  batch_data,
  penalty_mat = NULL,
  start = NULL,
  method = "nlminb",
  fit = TRUE,
  debug = 0,
  ...
)
}
\arguments{
\item{data}{A processed data frame produced by the function \code{\link{process_data}}}

\item{ddl}{A design data list produced by the function \code{\link{make_design_data}}.}

\item{model_parameters}{Model formula for the detection and movement portions
of the MMPP model. Must be of the form, e.g.,
\code{list(lambda=list(form=~1, offset=NULL), q=list(form=~1, offset=~log(1.num_neigh)))}. For the
\code{q} model you must use \code{offset=0} to not have one. If it is left off, \code{offset=~log(1.num_neigh)))}
will be used.}

\item{batch_data}{A data set with columns labeled \code{id} and \code{batch} that describe which
which batch each individual belongs to for the weighted-mean batch ML approximation of
Duncan (1980).}

\item{penalty_mat}{A covariance matrix for a multivariate normal penalty function. Using a
penalized likelihood helps with optimization. If left as \code{penalty_mat=NULL} then
no penalty is placed on the likelihood. If each batch contains sufficient observations
to estimate all parameters, then no penalty is necessary.}

\item{start}{Optional starting values for the parameter must be a list of the
form \code{list(beta_l=c(), beta_q=c())}.}

\item{method}{Optimization method. See \code{\link[optimx]{optimr}}}

\item{fit}{Logical. Should the likelihood be optimized?}

\item{debug}{Integer from 1-4. Opens browser() at various points in the function call. Mostly for
package developers.}

\item{...}{Additional arguments passed to the optimization function
\code{\link[optimx]{optimr}} from the \code{\link[optimx]{optimx-package}}.}
}
\description{
Fit MMPP movement model to resight data using a batch approximation to ML
}
\references{
Duncan, G. M. (1980). Approximate maximum likelihood estimation
with data sets that exceed computer limits.
Journal of Econometrics, 14(2), 257-264.
}
\author{
Devin S. Johnson
}
