% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design_data.R
\name{make_design_data}
\alias{make_design_data}
\title{Produce design data for use in fitting MMPP movement models}
\usage{
make_design_data(
  sighting_data,
  cell_data,
  add_fix = TRUE,
  add_boundary_length = FALSE
)
}
\arguments{
\item{sighting_data}{---.}

\item{cell_data}{---.}

\item{add_fix}{---.}

\item{add_boundary_length}{Logical. ---. Defaults to \code{FALSE}.}
}
\description{
Produce design data for use in fitting MMPP movement models
}
