% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sim_dm.R
\name{make_sim_dm}
\alias{make_sim_dm}
\title{Create model data necessary for MMPP simulation}
\usage{
make_sim_dm(ddl, model_parameters = mmpp_control())
}
\arguments{
\item{ddl}{Design data list created from \code{\link[moveMMPP]{make_design_data}}.}

\item{model_parameters}{Model formula for the detection and movement portions of the MMPP model.}
}
\description{
Create model data necessary for MMPP simulation
}
\author{
Devin S. Johnson
}
