% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sim_par.R
\name{make_sim_par}
\alias{make_sim_par}
\title{Create Q and L matrices for MMPP simulation}
\usage{
make_sim_par(sim_dm, par)
}
\arguments{
\item{sim_dm}{A list of design matrices resulting from a call to \code{\link[moveMMPP]{make_sim_dm}}}

\item{par}{A list corresponding to the coefficients for lambda, Q_m, and Q_r. It must match the
lengths in \code{sim_dm$par_map}.}
}
\description{
Create Q and L matrices for MMPP simulation
}
