% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process MMPP movement detection data for model fitting}
\usage{
process_data(data, aug_timestamp, time_unit = "days")
}
\arguments{
\item{data}{Detection data. Must contain columns \code{timestamp} and \code{cell} to
indicate time and place of detection. Also, must contain column named
\code{id} to indicate different individuals.}

\item{aug_timestamp}{Augmented times for modeling temporally dynamic movement or
detection. See 'Details'}

\item{time_unit}{---.}
}
\description{
Process MMPP movement detection data for model fitting
}
\details{
(\code{aug_timestamp} can be either passed as a separate vector of
POSIXct or \code{process_data} can derive a sequence of regularly spaced
augmentation times from the original data. This is specified by providing a
character string that corresponds to the by argument of the seq.POSIXt
function (e.g. '1 year', '1 month').
}
