% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_mmpp.R
\name{sim_mmpp}
\alias{sim_mmpp}
\title{Simulate a Markov Modulated Poisson Process}
\usage{
sim_mmpp(n = 1, max_time, min_obs, Q, L, start_loc = NULL)
}
\arguments{
\item{n}{Number of independent simulations}

\item{max_time}{...}

\item{min_obs}{...}

\item{Q}{A movement rate matrix}

\item{L}{A detection rate vector}

\item{start_loc}{Starting location of the path. If not specified, the limiting distribution
implied by \code{Q} will be used to randomly select a starting location.}
}
\description{
Simulate a Markov Modulated Poisson Process
}
\author{
Devin S. Johnson
}
