% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix.R
\name{dm_matrix}
\alias{dm_matrix}
\alias{dm_lambda}
\alias{dm_q_m}
\alias{dm_q_r}
\title{Create sparse representations of design matrices for
movement and resight models}
\usage{
dm_lambda(formula, ddl)

dm_q_m(formula, ddl)

dm_q_r(formula, ddl)
}
\arguments{
\item{ddl}{Design data list}

\item{par_list}{A list with named elements \code{form} ( the model formula) and
\code{offset} (formula for the offset; must result in a single column from call to \code{model.matrix})}
}
\description{
Create sparse representations of design matrices for
movement and resight models
}
\details{
This function is not designed for end-users but is exported for posterity.
}
