% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmpp.R
\name{fit_mmpp}
\alias{fit_mmpp}
\title{Fit MMPP movement model to resight data}
\usage{
fit_mmpp(
  data,
  ddl,
  model_parameters = list(lambda = ~1, q = ~1),
  fit_model = TRUE,
  start = NULL,
  ...
)
}
\arguments{
\item{data}{A processed data frame produced by the function \code{\link{process_data}}}

\item{ddl}{A design data list produced by the function \code{\link{make_design_data}}.}

\item{model_parameters}{Model formula for the detection and movement portions
of the MMPP model. Must be of the form \code{list(lambda=~form, q=~form)}.}

\item{fit_model}{Logical. Should the model be fit, or just return the TMB object?
Defaults to \code{fit_model = TRUE}.}

\item{start}{Optional starting values for the parameter must be a list of the
form \code{list(beta_l=c(), beta_q=c())}.}

\item{...}{Additional arguments passed to the optimization function
\code{\link[optimx]{opm}} from the \code{\link[optimx]{optimx-package}}.}
}
\description{
Fit MMPP movement model to resight data
}
\author{
Devin S. Johnson
}
