% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_functions.R
\name{get_lim_ud}
\alias{get_lim_ud}
\title{Get the limiting utilization distribution of the CTMC movement process}
\usage{
get_lim_ud(fit)
}
\arguments{
\item{fit}{A moveMMPP fitted model object from \code{\link[moveMMPP]{fit_mmpp_dir}} or
\code{\link[moveMMPP]{fit_mmpp_wmb}}.}
}
\description{
Get the limiting utilization distribution of the CTMC movement process
}
\author{
Devin S. Johnson
}
