% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveMMPP-package.R
\docType{package}
\name{moveMMPP-package}
\alias{moveMMPP}
\alias{moveMMPP-package}
\title{moveMMPP: Fit Continuous-Time Markov Modulated Poisson Process Movement Models to Animal Resight Data}
\description{
Using animal resight data, the package functions allow the user to estimate movement rates and resource selection using a continuous-time Markov chain movement model.
}
\author{
\strong{Maintainer}: Devin Johnson \email{devin.johnson@noaa.gov}

}
\keyword{internal}
